correct normalization process, add widefield planar example

+++ mmc/trunk/src/tettracing.h
@@ -69,7 +69,7 @@ typedef struct MMC_visitor{
 	int   reclen;
 	float *partialpath;
 	void  *photonseed;
-	double accumu_weight;
+	float accumu_weight;
 } visitor;
 
 #ifdef	__cplusplus
